backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
openai
google-genai
python-dotenv
websockets
